move ShortChannelParam
