+++ src/co/nayuta/lightning/Ptarmigan.java
@@ -71,45 +71,7 @@ public class Ptarmigan {
     private int peerFailCount = 0;
     private Logger logger;
 
-    public class ShortChannelParam {
-        int height;
-        int bIndex;
-        int vIndex;
-        byte[] minedHash;       //JNI戻り値専用
-        //
-        //
-        ShortChannelParam() {
-            this(-1, -1, -1);
-        }
-        ShortChannelParam(int height, int bIndex, int vIndex) {
-            this.height = height;
-            this.bIndex = bIndex;
-            this.vIndex = vIndex;
-            logger.debug("ShortChannelParam()=" + this.toString());
-        }
-        ShortChannelParam(long shortChannelId) {
-            initialize(shortChannelId);
-        }
-        //
-        void initialize(long shortChannelId) {
-            logger.debug("initialize()=" + shortChannelId);
-            if (shortChannelId != 0) {
-                this.height = (int)(shortChannelId >>> 40);
-                this.bIndex = (int)((shortChannelId & 0xffffff0000L) >>> 16);
-                this.vIndex = (int)(shortChannelId & 0xffff);
-            }
-            logger.debug("initialize()=" + this.toString());
-        }
-        //
-        boolean isAvailable() {
-            return (this.height > 0) && (this.bIndex > 0) && (this.vIndex >= 0);
-        }
-        //
-        @Override
-        public String toString() {
-            return String.format("height:%d, bIndex:%d, vIndex:%d", height, bIndex, vIndex);
-        }
-    }
+
     public static class SearchOutPointResult {
         int height;
         byte[] tx;
